Split out receiver and flight control
